# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Faux Api",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:5000/api/v1"
        }
    ],
    "paths": {
...
        "/courses": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get Courses List",
                "description": "Get data for all courses",
                "operationId": "courses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Data array",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Student ID",
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "description": "Course name",
                                                        "type": "string",
                                                        "example": "Math"
                                                    },
                                                    "description": {
                                                        "description": "Course short description",
                                                        "type": "string",
                                                        "example": "Something about numbers"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Wrong request!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Token not provided"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "action": {
                                                    "description": "name of the action where error is occurred",
                                                    "type": "string",
                                                    "example": "get_courses_list"
                                                },
                                                "message": {
                                                    "description": "Error message if incoming parameters is wrong",
                                                    "type": "string",
                                                    "example": "Fail due to some errors"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}